make it possible to mock a wallet connection in tests
