+++ deployables/app/app/wagmi/index.ts
@@ -1,0 +1,1 @@
+export { getWagmiConfig } from './getWagmiConfig'
