+++ deployables/app/package.json
@@ -44,7 +44,6 @@
     "zod": "3.24.1"
   },
   "devDependencies": {
-    "@depay/web3-mock": "14.19.1",
     "@moralisweb3/common-evm-utils": "^2.27.2",
     "@playwright/test": "1.50.1",
     "@react-router/dev": "7.1.5",
@@ -59,6 +58,7 @@
     "@types/react": "19.0.8",
     "@types/react-dom": "19.0.3",
     "@vitest/coverage-v8": "3.0.4",
+    "@wagmi/core": "^2.16.3",
     "@zodiac/eslint-config": "workspace:*",
     "@zodiac/test-utils": "workspace:*",
     "@zodiac/typescript-config": "workspace:*",
