+++ deployables/app/tsconfig.json
@@ -8,7 +8,8 @@
       "@/test-utils": ["./test-utils/index.ts"],
       "@/components": ["./app/components/index.ts"],
       "@/utils": ["./app/utils/index.ts"],
-      "@/balances": ["./app/balances/index.ts"]
+      "@/balances": ["./app/balances/index.ts"],
+      "@/wagmi": ["./app/wagmi/index.ts"]
     }
   },
   "include": [
